Rename files to fit naming convention
