+++ gs/frontend/aro/src/assets/orbital_logo.png
