+++ gs/frontend/aro/src/home/home_status.tsx
@@ -1,4 +1,4 @@
-import "./HomeStatus.css";
+import "./home_status.css";
 /**
  * @file HomeStatus.tsx
  * @brief Displays the status of the most recent request by a user or a message if there are no requests
