+++ gs/frontend/aro/src/new_request/new_request_form.tsx
@@ -1,4 +1,4 @@
-import "./NewRequestForm.css";
+import "./new_request_form.css";
 import { type ChangeEvent, useState } from "react";
 
 const NewRequestForm = () => {
