+++ gs/frontend/mcc/src/common/logs.tsx
