+++ gs/frontend/mcc/src/telemetry/telemetry_data.tsx
